face_api/requirements.txt
flask 
face_recognition 
numpy 
pillow 
flask_cors
python-dotenv